refactor(projective-grid): split grow.rs and topological/mod.rs at natural seams
Phase 6 file-size pass. Pure code motion, no behaviour change — bench
check is byte-compatible across all 27 public+private frames (pos/id/dup
counters zero) and the full test suite is green.

- seed_and_grow/grow.rs (992 lines) → grow/ directory:
  - grow/params.rs (220): SquareAttachPolicy trait + Admit /
    LabelledNeighbour / FillEdgeCtx data carriers + GrowParams /
    GrowResult.
  - grow/predict.rs (163): per-neighbour prediction geometry
    (collect_labelled_neighbours / predict_from_neighbours /
    is_extrapolating / local_step_at).
  - grow/mod.rs (655): BFS core, candidate search, boundary decision;
    re-exports the above so crate::seed_and_grow::grow::* paths are
    unchanged.

- topological/mod.rs (1140 lines): hex orchestration extracted to
  topological/hex_detect.rs (297). The square detection / merge / fit
  stay in mod.rs (876); shared helpers (triangulate_usable,
  ComponentOutput, MIN_USABLE_FOR_DELAUNAY) scoped pub(in crate::topological).
