+++ crates/projective-grid/src/seed_and_grow/grow/mod.rs
@@ -35,218 +35,30 @@
 //! collinearity / local-H residuals). See
 //! [`crate::shared::validate`](mod@crate::shared::validate) for
 //! that.
+//!
+//! # Module layout
+//!
+//! - `params` — the [`SquareAttachPolicy`] contract, the per-candidate
+//!   data carriers ([`Admit`], [`LabelledNeighbour`], [`FillEdgeCtx`]),
+//!   plus [`GrowParams`] / [`GrowResult`].
+//! - `predict` — the per-neighbour prediction geometry.
+//! - this module — the BFS core, candidate search / ambiguity resolution,
+//!   and boundary-cell decision, all re-exporting the above so existing
+//!   `crate::seed_and_grow::grow::*` paths are unchanged.
+
+mod params;
+mod predict;
+
+pub use params::{
+    Admit, FillEdgeCtx, GrowParams, GrowResult, LabelledNeighbour, SquareAttachPolicy,
+};
+pub(crate) use predict::{collect_labelled_neighbours, is_extrapolating, predict_from_neighbours};
 
 use kiddo::{KdTree, SquaredEuclidean};
 use nalgebra::{Point2, Vector2};
 use std::collections::{HashMap, HashSet, VecDeque};
 
 pub use crate::seed_and_grow::seed::Seed;
-
-/// Per-candidate decision from a [`SquareAttachPolicy`].
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum Admit {
-    /// Accept this candidate at the given grid cell.
-    Accept,
-    /// Reject this candidate; the generic code may move on to the
-    /// next nearest (if any).
-    Reject,
-}
-
-/// Information about an existing labelled neighbour, passed to the
-/// policy during candidate evaluation.
-#[derive(Clone, Copy, Debug)]
-pub struct LabelledNeighbour {
-    /// Index of the neighbour corner in the caller's position array.
-    pub idx: usize,
-    /// The neighbour's `(i, j)` grid cell.
-    pub at: (i32, i32),
-    /// The neighbour's position in image pixels.
-    pub position: Point2<f32>,
-}
-
-/// Caller-supplied attachment policy for [`bfs_grow`].
-///
-/// Implementations typically hold references to the caller's feature
-/// data (axes, labels, strengths) plus tuning parameters, and use `idx`
-/// to look up the relevant per-feature record inside each callback.
-pub trait SquareAttachPolicy {
-    /// Is this corner index a possible candidate at all? Called
-    /// once per corner when the KD-tree is built.
-    fn is_eligible(&self, idx: usize) -> bool;
-
-    /// Optional caller-defined label required at grid cell `(i, j)`.
-    /// Return `None` for no constraint.
-    fn required_label_at(&self, i: i32, j: i32) -> Option<u8>;
-
-    /// Return the label of the corner at `idx`. Must agree with
-    /// `required_label_at` at attachment time. Called during
-    /// candidate filtering.
-    fn label_of(&self, idx: usize) -> Option<u8>;
-
-    /// Accept or reject a candidate for attachment at grid cell
-    /// `at` given its geometric prediction and existing labelled
-    /// neighbours. Called per candidate in order of increasing
-    /// distance to `prediction`.
-    fn accept_candidate(
-        &self,
-        idx: usize,
-        at: (i32, i32),
-        prediction: Point2<f32>,
-        neighbours: &[LabelledNeighbour],
-    ) -> Admit;
-
-    /// Soft per-edge check: is the induced edge between the just-
-    /// attached candidate and one of its cardinal-labelled neighbours
-    /// admissible? At least one cardinal edge must pass for the
-    /// attachment to stick; otherwise the position is marked a hole
-    /// and the candidate is rolled back.
-    ///
-    /// Default: accept all edges (no soft check).
-    fn edge_ok(
-        &self,
-        _candidate_idx: usize,
-        _neighbour_idx: usize,
-        _at_candidate: (i32, i32),
-        _at_neighbour: (i32, i32),
-    ) -> bool {
-        true
-    }
-
-    /// Optional widened eligibility used by the fill-pass booster.
-    ///
-    /// Defaults to [`Self::is_eligible`]; patterns whose precision
-    /// core admits only `Clustered` corners but want to admit a few
-    /// near-cluster corners during the booster pass override this to
-    /// expand the admissible set. The fill pass calls this when
-    /// building its KD-tree; the regular grow / boundary-extension
-    /// passes ignore it.
-    fn eligible_for_fill(&self, idx: usize) -> bool {
-        self.is_eligible(idx)
-    }
-
-    /// Optional fill-pass edge check that has access to the full
-    /// labelled set and the position table via [`FillEdgeCtx`].
-    ///
-    /// The default delegates to [`Self::edge_ok`], ignoring the extra
-    /// context. Pattern implementations that need a directional edge
-    /// metric (e.g., a strongly anisotropic component where the
-    /// horizontal pitch is much larger than the vertical pitch and a
-    /// scalar `cell_size` rejects legitimate vertical extrapolations)
-    /// override this to consult the labelled set when computing the
-    /// expected edge length.
-    ///
-    /// Only invoked by [`crate::seed_and_grow::fill::fill_grid_holes`]; the
-    /// regular grow and boundary-extension passes call [`Self::edge_ok`]
-    /// directly.
-    fn fill_edge_ok(&self, ctx: FillEdgeCtx<'_>) -> bool {
-        self.edge_ok(
-            ctx.candidate_idx,
-            ctx.neighbour_idx,
-            ctx.at_candidate,
-            ctx.at_neighbour,
-        )
-    }
-}
-
-/// Context passed to [`SquareAttachPolicy::fill_edge_ok`].
-///
-/// Bundles every piece of state the policy needs to make a
-/// labelled-set-aware edge decision: the candidate + cardinal
-/// neighbour indices, their `(i, j)` cells, the full labelled map,
-/// the corner position array, and the scalar fallback cell size.
-#[non_exhaustive]
-#[derive(Clone, Copy)]
-pub struct FillEdgeCtx<'a> {
-    /// Index of the candidate corner being evaluated.
-    pub candidate_idx: usize,
-    /// Index of the already-labelled cardinal neighbour.
-    pub neighbour_idx: usize,
-    /// The candidate's prospective `(i, j)` cell.
-    pub at_candidate: (i32, i32),
-    /// The cardinal neighbour's `(i, j)` cell.
-    pub at_neighbour: (i32, i32),
-    /// The full `(i, j) → corner_idx` labelled map at this point in the grow.
-    pub labelled: &'a HashMap<(i32, i32), usize>,
-    /// Corner positions in image pixels, indexed by the values of `labelled`.
-    pub positions: &'a [Point2<f32>],
-    /// Scalar fallback cell size in pixels, used when no local estimate exists.
-    pub cell_size: f32,
-}
-
-/// Tolerances for [`bfs_grow`].
-#[non_exhaustive]
-#[derive(Clone, Copy, Debug)]
-pub struct GrowParams {
-    /// Candidate-search radius (fraction of `cell_size`) around each
-    /// prediction. Applies when the target is being **interpolated**
-    /// between labelled neighbours on opposite sides.
-    pub attach_search_rel: f32,
-    /// Ambiguity factor: if the second-nearest candidate is within
-    /// `factor × nearest_distance`, the attachment is skipped.
-    pub attach_ambiguity_factor: f32,
-    /// Multiplier on `attach_search_rel` when the target is being
-    /// **extrapolated** outward from the labelled set (every labelled
-    /// neighbour sits on the same side of the target along at least one
-    /// axis). Defaults to 2.0 — opens the search up enough to absorb
-    /// the perspective-foreshortening overshoot at the image edge while
-    /// still rejecting off-lattice candidates that sit several cell-
-    /// widths away.
-    pub boundary_search_factor: f32,
-}
-
-impl Default for GrowParams {
-    fn default() -> Self {
-        Self {
-            attach_search_rel: 0.35,
-            attach_ambiguity_factor: 1.5,
-            boundary_search_factor: 2.0,
-        }
-    }
-}
-
-impl GrowParams {
-    /// Construct grow parameters from the interpolation search radius and
-    /// ambiguity factor; `boundary_search_factor` keeps its default.
-    pub fn new(attach_search_rel: f32, attach_ambiguity_factor: f32) -> Self {
-        Self {
-            attach_search_rel,
-            attach_ambiguity_factor,
-            ..Self::default()
-        }
-    }
-}
-
-/// Outcome of a grow pass.
-#[derive(Debug, Default)]
-pub struct GrowResult {
-    /// `(i, j) → corner_index` map of accepted labels. Rebased so the
-    /// bounding-box minimum is `(0, 0)`.
-    pub labelled: HashMap<(i32, i32), usize>,
-    /// Inverse map.
-    pub by_corner: HashMap<usize, (i32, i32)>,
-    /// Positions with ≥ 2 candidates inside the ambiguity window.
-    pub ambiguous: HashSet<(i32, i32)>,
-    /// Positions with no accepted candidate.
-    pub holes: HashSet<(i32, i32)>,
-    /// Grid `i`-axis vector (pixels per cell) carried forward — overlays
-    /// and boosters use it.
-    pub axis_i: Vector2<f32>,
-    /// Grid `j`-axis vector (pixels per cell) carried forward — overlays
-    /// and boosters use it.
-    pub axis_j: Vector2<f32>,
-    /// Mod-2 `i` component of the coordinate shift removed by the
-    /// final rebase.
-    ///
-    /// `bfs_grow` walks in seed-local coordinates, then subtracts the
-    /// labelled bounding-box minimum so output coordinates start at
-    /// `(0, 0)`. Callers with an alternating label rule can add these
-    /// mod-2 components back when evaluating labels in post-rebase
-    /// coordinates. Callers without an alternating rule can ignore
-    /// these fields.
-    pub rebase_i_mod2: i32,
-    /// See [`Self::rebase_i_mod2`].
-    pub rebase_j_mod2: i32,
-}
 
 /// Grow a labelled `(i, j)` grid from a 2×2 seed using BFS over the
 /// lattice boundary.
@@ -400,154 +212,6 @@ pub(super) fn enqueue_cardinal_neighbours(
         if !labelled.contains_key(&neigh) && seen.insert(neigh) {
             boundary.push_back(neigh);
         }
-    }
-}
-
-pub(crate) fn collect_labelled_neighbours(
-    pos: (i32, i32),
-    window_half: i32,
-    labelled: &HashMap<(i32, i32), usize>,
-    positions: &[Point2<f32>],
-) -> Vec<LabelledNeighbour> {
-    let mut out = Vec::new();
-    for dj in -window_half..=window_half {
-        for di in -window_half..=window_half {
-            if di == 0 && dj == 0 {
-                continue;
-            }
-            let at = (pos.0 + di, pos.1 + dj);
-            if let Some(&idx) = labelled.get(&at) {
-                out.push(LabelledNeighbour {
-                    idx,
-                    at,
-                    position: positions[idx],
-                });
-            }
-        }
-    }
-    out
-}
-
-/// Distance-weighted average of per-neighbour axis-vector predictions.
-///
-/// Use this function for in-the-loop BFS attachment where arbitrary
-/// labelled neighbours are available.
-///
-/// For each labelled neighbour `N_k` at `(i_k, j_k)`, the prediction is
-/// `pred_k = pos(N_k) + (Δi · i_step_k) + (Δj · j_step_k)` where
-/// `Δi = target.i − i_k`, `Δj = target.j − j_k`, and `i_step_k` /
-/// `j_step_k` are the **local** grid-step vectors observed at `N_k`:
-///
-/// - If `(i_k+1, j_k)` and `(i_k−1, j_k)` are both labelled, the i-step is
-///   the central difference `(pos(i_k+1, j_k) − pos(i_k−1, j_k)) / 2`.
-/// - Otherwise, a one-sided difference from whichever neighbour is
-///   labelled.
-/// - Otherwise, fall back to the global `cell_size · u`. Same for j.
-///
-/// This linearises the grid **at every neighbour individually** instead of
-/// trusting the seed's global `(u, v, cell_size)` — critical under strong
-/// perspective foreshortening, where the cell pitch on the far edge of
-/// the labelled set is materially different from the seed's mean. With
-/// the global-only model, BFS predictions on the foreshortened side
-/// overshoot the next true corner by more than the search radius and
-/// growth terminates prematurely.
-///
-/// Predictions are averaged with weights `1 / (Δi² + Δj²)` so cardinal
-/// neighbours (grid distance 1) carry weight 1.0 while diagonal
-/// neighbours (grid distance √2) carry weight 0.5 — variance addition
-/// per grid step.
-///
-/// A neighbour at the target cell itself (`Δi = Δj = 0`) would yield an
-/// infinite weight; in practice [`bfs_grow`] never enqueues such a
-/// neighbour (they're already labelled), but for robustness we treat
-/// `Δi = Δj = 0` as weight 1.0 to avoid `NaN`.
-pub(crate) fn predict_from_neighbours(
-    target: (i32, i32),
-    neighbours: &[LabelledNeighbour],
-    u: Vector2<f32>,
-    v: Vector2<f32>,
-    cell_size: f32,
-    labelled: &HashMap<(i32, i32), usize>,
-    positions: &[Point2<f32>],
-) -> Point2<f32> {
-    debug_assert!(!neighbours.is_empty());
-    let global_i_step = u * cell_size;
-    let global_j_step = v * cell_size;
-
-    let mut sum_x = 0.0_f32;
-    let mut sum_y = 0.0_f32;
-    let mut sum_w = 0.0_f32;
-    for n in neighbours {
-        let di = (target.0 - n.at.0) as f32;
-        let dj = (target.1 - n.at.1) as f32;
-        let d2 = di * di + dj * dj;
-        let w = if d2 > 0.0 { 1.0 / d2 } else { 1.0 };
-
-        let i_step = local_step_at(n.at, (1, 0), labelled, positions).unwrap_or(global_i_step);
-        let j_step = local_step_at(n.at, (0, 1), labelled, positions).unwrap_or(global_j_step);
-
-        let off = i_step * di + j_step * dj;
-        sum_x += w * (n.position.x + off.x);
-        sum_y += w * (n.position.y + off.y);
-        sum_w += w;
-    }
-    Point2::new(sum_x / sum_w, sum_y / sum_w)
-}
-
-/// True when every labelled neighbour sits on the same side of `target`
-/// along at least one of the two grid axes — i.e., the target is being
-/// extrapolated outward from the labelled set rather than interpolated
-/// between two opposing sides.
-///
-/// This is the geometric signal that the search prediction is less
-/// reliable: extrapolation accumulates foreshortening error linearly,
-/// while interpolation has neighbours on both sides bracketing the
-/// truth.
-pub(super) fn is_extrapolating(target: (i32, i32), neighbours: &[LabelledNeighbour]) -> bool {
-    let mut has_neg_di = false;
-    let mut has_pos_di = false;
-    let mut has_neg_dj = false;
-    let mut has_pos_dj = false;
-    for n in neighbours {
-        let di = target.0 - n.at.0;
-        let dj = target.1 - n.at.1;
-        if di > 0 {
-            has_neg_di = true; // neighbour is on the −i side of target
-        } else if di < 0 {
-            has_pos_di = true;
-        }
-        if dj > 0 {
-            has_neg_dj = true;
-        } else if dj < 0 {
-            has_pos_dj = true;
-        }
-    }
-    !(has_neg_di && has_pos_di && has_neg_dj && has_pos_dj)
-}
-
-/// Estimate the local grid-step vector at labelled cell `at` along
-/// direction `step = (di, dj)` using a finite-difference of labelled
-/// neighbours. Returns `None` when neither the forward nor backward
-/// neighbour is labelled.
-pub(super) fn local_step_at(
-    at: (i32, i32),
-    step: (i32, i32),
-    labelled: &HashMap<(i32, i32), usize>,
-    positions: &[Point2<f32>],
-) -> Option<Vector2<f32>> {
-    let here = labelled.get(&at).map(|&i| positions[i])?;
-    let fwd = (at.0 + step.0, at.1 + step.1);
-    let bwd = (at.0 - step.0, at.1 - step.1);
-    let fwd_pos = labelled.get(&fwd).map(|&i| positions[i]);
-    let bwd_pos = labelled.get(&bwd).map(|&i| positions[i]);
-    match (fwd_pos, bwd_pos) {
-        (Some(f), Some(b)) => {
-            let v = (f - b) * 0.5;
-            Some(v)
-        }
-        (Some(f), None) => Some(f - here),
-        (None, Some(b)) => Some(here - b),
-        (None, None) => None,
     }
 }
